src/answer_processing/pyproject.toml
[project]
name = "ocr-asr-answer-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fugashi>=1.5.2",
    "numpy==1.26.4",
    "pillow>=12.0.0",
    "protobuf>=6.33.1",
    "sentencepiece>=0.2.1",
    "soundfile>=0.13.1",
    "torch==2.2",
    "transformers>=4.57.1",
    "unidic-lite>=1.0.8",
]
